invoice_uploader/requirements.txt
boto3==1.34.0 
kafka-python==2.0.2 
python-dotenv==1.0.1 
psycopg2-binary==2.9.9